# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Reset Pixel Art Tool API",
    "version": "1.0.0",
    "description": "API for users and pixel art file operations. Login returns a JWT token and protected routes require Authorization: Bearer <token>."
  },
  "servers": [{ "url": "http://localhost:3000" }],
  "paths": {
    "/api/users": {
      "get": {
        "summary": "List users",
        "responses": { "200": { "description": "Array of users" } }
      },
      "post": {
        "summary": "Create user",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NewUser" } } } },
        "responses": { "201": { "description": "Created" }, "400": { "description": "Bad request" } }
      }
    },
    "/api/users/login": {
      "post": {
        "summary": "Login user",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Login" } } } },
        "responses": { "200": { "description": "User info and JWT token" }, "401": { "description": "Unauthorized" } }
      }
    },
    "/api/users/{userName}": {
      "get": { "summary": "Get user", "parameters": [{ "name": "userName", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "User" }, "404": { "description": "Not found" } } },
      "put": { "summary": "Update user", "parameters": [{ "name": "userName", "in": "path", "required": true, "schema": { "type": "string" } }], "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateUser" } } } }, "responses": { "200": { "description": "Updated" } } },
      "delete": { "summary": "Delete user", "parameters": [{ "name": "userName", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Deleted" }, "404": { "description": "Not found" } } }
    },
    "/api/files/create": {
      "post": {
        "summary": "Create file",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateFile" } } } },
        "responses": { "201": { "description": "Created" }, "400": { "description": "Bad request" } }
      }
    },
    "/api/files": {
      "get": { "summary": "List files for user", "responses": { "200": { "description": "File list" } } }
    },
    "/api/files/{fileName}": {
      "get": { "summary": "Get file state", "parameters": [{ "name": "fileName", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "File" }, "404": { "description": "Not found" } } },
      "delete": { "summary": "Delete file", "parameters": [{ "name": "fileName", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Deleted" } } }
    },
    "/api/files/{fileName}/palette/add": {
      "post": { "summary": "Add palette color", "responses": { "200": { "description": "OK" }, "400": { "description": "Bad request" } } }
    },
    "/api/files/{fileName}/palette/{colorIndex}": {
      "put": { "summary": "Update palette color", "responses": { "200": { "description": "OK" } } }
    },
    "/api/files/{fileName}/pixel": { "put": { "summary": "Recolor a pixel", "responses": { "200": { "description": "OK" } } } },
    "/api/files/{fileName}/undo": { "post": { "summary": "Undo action", "responses": { "200": { "description": "OK" } } } },
    "/api/files/{fileName}/redo": { "post": { "summary": "Redo action", "responses": { "200": { "description": "OK" } } } }
  },
  "components": {
    "schemas": {
      "User": { "type": "object", "properties": { "userName": { "type": "string" }, "email": { "type": "string" }, "files": { "type": "array", "items": { "type": "string" } } } },
      "NewUser": { "type": "object", "required": ["userName","email","password"], "properties": { "userName": { "type": "string" }, "email": { "type": "string" }, "password": { "type": "string" } } },
      "Login": { "type": "object", "required": ["userName","password"], "properties": { "userName": { "type": "string" }, "password": { "type": "string" } } },
      "UpdateUser": { "type": "object", "properties": { "email": { "type": "string" }, "password": { "type": "string" }, "files": { "type": "array", "items": { "type": "string" } } } },
      "CreateFile": { "type": "object", "required": ["fileName","width","height"], "properties": { "fileName": { "type": "string" }, "width": { "type": "integer" }, "height": { "type": "integer" } } }
    }
  }
}